Refactor CLI and GUI clients to use a common interface
This change refactors the command-line and graphical user interface clients to share a common architecture.

The key changes include:
- A new `IClient` abstract base class that defines the common interface for all clients.
- A new `BaseClient` class that encapsulates the shared logic of creating and managing the `CommandProcessor` and `ResultRepository`.
- New `ConsoleClient` and `GuiClient` classes that inherit from `BaseClient` and contain the application-specific logic for the CLI and GUI, respectively.
- The main entry points (`CLImain.cpp` and `GUImain.cpp`) have been simplified to just create and run their corresponding client objects.
- The root `CMakeLists.txt` has been updated to build a new `ClientLib` containing the refactored client code and link the executables against it.
