+++ include/MITSUDomoe/ConsoleClient.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include "BaseClient.hpp"
+
+namespace MITSU_Domoe
+{
+
+class ConsoleClient : public BaseClient
+{
+public:
+    ConsoleClient();
+    ~ConsoleClient() = default;
+
+    void run() override;
+};
+
+}
